reverse nightlies list
Will fix branch version selection for changelogs.
When a device is transitioned to a newer branch (eg 14.1 to 15.1), the changelogs continue to show the old branch, because the first nightly in a device's nightly list is used to decide the branch version (line 135).
Now that the nightlies list is reversed, the newest nightly is first, so the latest branch will be chosen