updater: make device json file paths configurable
In production we run the updater in docker, and attach the two device
json files via volumes. Because of the way Docker's volumes work, if
the original file is overwritten and its inode number changes, the
modification isn't reflected inside the container.

To work around this, make the json paths configurable so we can put them
in their own directory and avoid the issue altogether.